refactor swaps example to incoming and outgoing
